Modified BME280 code to communicate with BMP280.
As I bought a new, integrated board that contains MPU9250 and BMP280
sensors together, I needed to make minor changes to the code so that it
is able to communicate with BMP280. BME280 is downward compatible with
BMP280, so this is possible.

Altitude measurements drifts a lot over time, hence I increased the
calibration period quite a bit (10000 samples are taken, each with 1ms
delay).